started making enemy class

+++ core/src/main/java/io/github/team9/escapefromuni/Coordinate.java
@@ -1,24 +1,24 @@
 package io.github.team9.escapefromuni;
 
 public class Coordinate {
-    public int x;
-    public int y;
-    public Coordinate(int x, int y) {
+    public float x;
+    public float y;
+    public Coordinate(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
